requirements.txt
numpy
scikit-learn
joblib
google-cloud-firestore
